refactor🎨: SCSS 变量改由 module 文件导出以适配 Vite
