+++ src/layout/components/Sidebar/Logo.vue
@@ -15,7 +15,7 @@
 
 <script>
 
-import variables from '@/styles/variables.scss'
+import variables from '@/styles/variables.module.scss'
 import { mapGetters } from 'vuex'
 
 export default {
