+++ src/layout/components/Sidebar/index.vue
@@ -29,7 +29,7 @@
 import { mapGetters } from 'vuex'
 import Logo from './Logo'
 import SidebarItem from './SidebarItem'
-import variables from '@/styles/variables.scss'
+import variables from '@/styles/variables.module.scss'
 
 export default {
   components: { SidebarItem, Logo },
